ARIES-1421 SimpleFilter attribute extraction can not handle version ranges
Add test.

Refactor SimpleFilter.attributes so that it returns a map of attribute names each associate with a list of simple filters. Each simple filter
in the list represents one occurence of the attribute in the filter. Each simple filter in the list is guaranteed to be either of the form
(attribute operation value) or (!(attribute operation value)). The caller is responsible for sorting out the semantics.

Add new method to the abstract header class capable of parsing a version range from a list of simple filters of the form described above.

git-svn-id: https://svn.apache.org/repos/asf/aries/trunk@1705502 13f79535-47bb-0310-9956-ffa450edef68