Region Name Test
Add a small test to enforce the current subsystem region name convention, which is subsystemSymbolicName;subsystemVersion;subsystemType;subsystemId. At least one major Aries
Subsystems consumer has a business requirement relying on it. The idea is to prevent an unannounced or inadvertent change without notification or discussion. The exact form of
the pattern is less important than it being known and predictable.

git-svn-id: https://svn.apache.org/repos/asf/aries/trunk@1707545 13f79535-47bb-0310-9956-ffa450edef68